# proc: only format registers value when it's necessary (#1860)